gopls/internal/util/lru: Move cache to its own internal utility in internal/lru
This change exposes the LRU cache for use in other modules, such as internal/imports,
which may use it in the future to support caching across repeated invocations.
